+++ packages/viking-ui/src/public-api.ts
@@ -151,6 +151,7 @@ export * from "./lib/kanban/kanban";
 export * from "./lib/kbd/kbd";
 export * from "./lib/label/label";
 export * from "./lib/loading-overlay/loading-overlay";
+export * from "./lib/layout/layout";
 export * from "./lib/metric-card/metric-card";
 export * from "./lib/menubar/menubar";
 export * from "./lib/modal/modal";